+++ src/replayx/__init__.py
@@ -9,7 +9,7 @@ See https://github.com/mkusiappiah/replayx for documentation.
 
 from __future__ import annotations
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 from ._types import RecordMode
 from .cassette import Cassette, Interaction, RecordedRequest, RecordedResponse